Fixed bug where a blank card is displayed
on-tap is attached to both the chip-container and the card.
To know which card gets displayed, the transition function called
by on-tap assigns the selectedAlbum variable the value of the target's
templateInstance.model.item, which is null if the user clicks on any
area which is inside chip-container but outside a chip. This results
to a blank card being displayed.

The simplest fix for this is to check if selectedAlbum is null before
setting the page flag to 1.